Cron task fix
Currently, if the connection/command somewhy dropped out or some other
internet issue occurred, the server would get rebooted even tho it was
actually up. Added a secondary check to the server checker. When the
screen is down / no output received, it will wait for 3 seconds and fire
the checker again. If it still doesn't get any output, it restarts the
server.